+++ packages/plugin-divider/web/src/typeMapper.ts
@@ -1,4 +1,4 @@
-import DividerComponent from './components/divider-component';
+import DividerComponent from './divider-component';
 import { DIVIDER_TYPE } from './types';
 import { PluginTypeMapper } from 'wix-rich-content-common';
 import { sizeClassName } from './classNameStrategies';